+++ 4-semester-java/3-movie-service/movies/src/main/java/pl/pjatk/sansob/movies/services/MovieService.java
@@ -7,7 +7,6 @@ import pl.pjatk.sansob.movies.models.Movie;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Service
 public class MovieService {
@@ -16,14 +15,14 @@ public class MovieService {
 
     public MovieService() {
         this.movies = new ArrayList<>();
-        this.movies.add(new Movie(1, "Harry Potter", Category.Fantasy));
+        this.movies.add(new Movie(1L, "Harry Potter", Category.Fantasy));
     }
 
     public List<Movie> getAll() {
         return this.movies;
     }
 
-    public Optional<Movie> getById(Integer id) {
+    public Optional<Movie> getById(Long id) {
         return this.movies
                 .stream()
                 .filter(c -> c.getId().equals(id))
